[Quantization][NFC] Move some structs/using to QuantizationBase
These make more sense in `QuantizationBase` and allow other files using them to include less.

This included changing the constructor for `NodeNameAndKind` to just take the info from the `NodeValue` instead of the `NodeValue` itself, so that it doesn't need to depend on anything from Graph.
