+++ tests/unittests/BackendTestUtils.h
@@ -17,14 +17,13 @@
 #include "glow/Backends/Backend.h"
 #include "glow/ExecutionEngine/ExecutionEngine.h"
 #include "glow/Graph/Graph.h"
-#include "glow/IR/IR.h"
-
-#include "llvm/Support/Casting.h"
-
-#include "glow/Backends/Backend.h"
 #include "glow/Graph/Node.h"
+#include "glow/IR/IR.h"
 #include "glow/IR/IRBuilder.h"
 #include "glow/IR/IRGen.h"
+#include "glow/Quantization/Base/Base.h"
+
+#include "llvm/Support/Casting.h"
 
 #include "gtest/gtest.h"
